cdc/requirements.txt
mysql-replication==1.0.9
packaging==24.2
PyMySQL==1.1.1
cryptography

